backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
motor
beanie
python-dotenv
httpx
pymongo[srv]
dnspython
requests
